Remove unnecessary `static_cast`s

+++ src/CSFML/Graphics/Shader.cpp
@@ -253,10 +253,10 @@ void sfShader_setIntColorUniform(sfShader* shader, const char* name, sfColor col
     sfShader_setIvec4Uniform(shader,
                              name,
                              {
-                                 static_cast<int>(color.r),
-                                 static_cast<int>(color.g),
-                                 static_cast<int>(color.b),
-                                 static_cast<int>(color.a),
+                                 color.r,
+                                 color.g,
+                                 color.b,
+                                 color.a,
                              });
 }
